RestApiProducts/requirements.txt
Django==4.1.7
asgiref==3.6.0
sqlparse==0.4.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
PyJWT==2.6.0
requests==2.28.2